% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfa_evar_near_zero}
\alias{cfa_evar_near_zero}
\title{Generated from a two-factor model, with one error variacne close to zero.}
\format{
An object of class \code{data.frame} with 120 rows and 6 columns.
}
\usage{
cfa_evar_near_zero
}
\description{
Generated from a two-factor model, with one error variacne close to zero.
}
\keyword{datasets}
