% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfa_two_factors}
\alias{cfa_two_factors}
\title{Dataset: Two factors, six variables, 500 cases}
\format{
A data framw with 500 rows and six variables.
}
\usage{
cfa_two_factors
}
\description{
Generated from a two-factor model with six variables, n = 500
}
\examples{
\dontrun{

 library(lavaan)
 mod <- "f1 =~ x1 + x2 + x3
         f2 =~ x4 + x5 + x6"
 fit <- cfa(mod, cfa_two_factors)
 summary(fit)
}
}
\keyword{datasets}
