% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mg.R
\docType{data}
\name{cfa_two_factors_mg}
\alias{cfa_two_factors_mg}
\title{Dataset (CFA, Two Factors, Six Variables, Two Groups)}
\format{
A data frame with 500 rows, one grouping variable, \code{gp},
six variables, \code{x1} to \code{x6}.
}
\usage{
cfa_two_factors_mg
}
\description{
Generated from a two-factor model with six variables, n = 500,
two groups, n = 250 each.
}
\details{
Supposed to be tested using this example:\preformatted{library(lavaan)
mod <- "f1 =~ x1 + x2 + x3
        f2 =~ x4 + x5 + x6"
fit <- cfa(mod, cfa_two_factors_mg, group = "gp")
summary(fit)
}
}
\keyword{datasets}
