% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sem_out.R
\name{check_sem_out}
\alias{check_sem_out}
\title{Check the SEM output in semlbci}
\usage{
check_sem_out(sem_out, robust = "none")
}
\arguments{
\item{sem_out}{The output from an SEM analysis. Currently only
supports a \link[lavaan:lavaan-class]{lavaan::lavaan} object.}

\item{robust}{Whether the LBCI based on robust likelihood ratio test is to
be found. Only "satorra.2000" in \link{lavaan} is supported for
now. If "none", the default, then likelihood ratio test based
on maximum likelihood estimation will be used.}
}
\value{
A numeric vector of one element. If 0, the model and
estimation method are officially supported. If larger than zero,
then the model and method are not officially supported. If less
than zero, then the model and/or the method are officially
not supported.

The attributes \code{info} contains the reason for a value other
than zero.
}
\description{
Check whether the model and the estimation method in the
\code{sem_out}
object passed to \code{\link[=semlbci]{semlbci()}}
are supported by the current version of
\code{\link[=semlbci]{semlbci()}}.
}
\details{
This function is to be used by \code{\link[=semlbci]{semlbci()}} but is exported such that
the compatibility of an SEM output can be checked independently.

Estimation methods (\code{estimator} in
\code{\link[lavaan:lavaan]{lavaan::lavaan()}}) currently supported:
\itemize{
\item Maximum likelihood (\code{ML})
\item Full information maximum likelihood with missing data (\code{fiml})
\item Generalized least squares (\code{GLS})
\item Weighted least squares (a.k.a.
asymptotically distribution free) (\code{WLS})
}

Estimation methods not yet supported:
\itemize{
\item Unweighted least squares (\code{ULS})
\item Diagonally weighted least squares (\code{DWLS})
\item Variants with robust standard errors and/or robust
test statistics:
\itemize{
\item \code{MLM}, \code{MLMV}, \code{MLMVS}, \code{MLF}, and \code{MLR}.
\item \code{WLSM}, \code{WLSMV}.
\item \code{ULSM}, \code{ULSMV}.
}
}

Currently supported models:
\itemize{
\item Single group models with continuous variables.
}

Models not tested:
\itemize{
\item Models with categorical variables.
}

Models not yet supported:
\itemize{
\item Multigroup models.
\item Models with formative factors
\item Multilevel models
}
}
\examples{
library(lavaan)
data(cfa_two_factors)
mod <-
"
f1 =~ x1 + x2 + x3
f2 =~ x4 + x5 + x6
"

fit <- sem(mod, cfa_two_factors)

# Should be 0
check_sem_out(fit)

fit2 <- sem(mod, cfa_two_factors, estimator = "MLR")

# Should be negative because MLR is officially not supported
check_sem_out(fit2)

}
\seealso{
\code{\link[=semlbci]{semlbci()}}, \code{\link[=ci_i]{ci_i()}}
}
