% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_i.R
\name{ci_i}
\alias{ci_i}
\title{Find the lower bound and upper bound for one parameter}
\usage{
ci_i(i, ...)
}
\arguments{
\item{i}{The position of the target parameters.}

\item{...}{Arguments to be passed to \code{ci_bound_i}.}
}
\value{
The requested bounds
Can return the optimization history as an attribute.
}
\description{
Find the lower and upper bound for one parameter.
}
\details{
Currently supports \code{lavaan} output only.
}
\examples{
library(lavaan)
data(cfa_two_factors)
mod <-
"
f1 =~ x1 + x2 + a*x3
f2 =~ x4 + a*x5 + equal('f1=~x2')*x6
f1 ~~ 0*f2
asq := a^2
"
fit <- sem(mod, cfa_two_factors)
}
