% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_i_one.R
\name{ci_i_one}
\alias{ci_i_one}
\title{LB Confidence Limit for One Parameter}
\usage{
ci_i_one(
  i,
  which = NULL,
  sem_out,
  method = "wn",
  standardized = FALSE,
  robust = "none",
  sf_full = NA,
  sf_args = list(),
  sem_out_name = NULL,
  try_k_more_times = 0,
  ...
)
}
\arguments{
\item{i}{The position of the target parameters as appeared in the
parameter table of the \link[lavaan:lavaan-class]{lavaan::lavaan} object.}

\item{which}{Whether the lower bound or the upper bound is to be
found. Must be \code{"lbound"} or \code{"ubound"}.}

\item{sem_out}{The SEM output. Currently supports
\link[lavaan:lavaan-class]{lavaan::lavaan} outputs only.}

\item{method}{The approach to be used. Default is "wn"
(Wu-Neale-2012 Method). The other methods are disabled for now.}

\item{standardized}{Boolean. Whether the LBCI for the standardized
solution is to be searched. Default is \code{FALSE}.}

\item{robust}{Whether the LBCI based on robust likelihood ratio
test is to be found. Only "satorra.2000" in \link{lavaan} is supported
for now. If `"none"``, the default, then likelihood ratio test based
on maximum likelihood estimation will be used.}

\item{sf_full}{A list with the scaling factors. Ignored if \code{robust}
is \code{"none"}. If \code{robust} is \code{"satorra.2000"} and \code{sf_full} is
supplied, then its value will be used.
If \code{robust} is \code{"satorra.2000"} but \code{sf_full} is
\code{NA}, then scaling factors will be computed internally.}

\item{sf_args}{The list of arguments to be used for computing scaling factors
if \code{robust} is \code{"satorra.2000"}. Used only by \code{\link[=semlbci]{semlbci()}}.}

\item{sem_out_name}{The name of the object supplied to \code{sem_out}. \code{NULL}
by default. To be used by internal functions.}

\item{try_k_more_times}{How many more times to try if the status code is not zero.
Default is 0.}

\item{...}{Arguments to be passed to the function corresponds to
the requested method (\code{\link[=ci_bound_wn_i]{ci_bound_wn_i()}} for "wn").}
}
\value{
A numeric vector of two elements. The first element is the
lower bound, and the second element is the upper bound.

The diagnostic information from the function called in finding the
lower and upper founds are stored in the attributes \code{lb_diag} and
\code{ub_diag}, for the lower bound and the upper bound, respectively.
}
\description{
Finds the likelihood-based confidence limit
for one parameter.
}
\details{
\code{\link[=ci_i_one]{ci_i_one()}} calls a function (\code{\link[=ci_bound_wn_i]{ci_bound_wn_i()}} by default)
once to find the bound (limit) for a confidence interval.
The default method is the Wu-Neale-2012 method. Please refer to
\code{\link[=ci_bound_wn_i]{ci_bound_wn_i()}} for further information.

This function is not supposed to be used directly by users. It is
exported such that interested users can examine how a confidence bound it
found.
}
\examples{

data(simple_med)

library(lavaan)
mod <-
"
m ~ x
y ~ m
"
fit_med <- lavaan::sem(mod, simple_med, fixed.x = FALSE)

parameterTable(fit_med)

# Find the LBCI for the first parameter
# THe method "wn" need the constraint function:
fn_constr0 <- set_constraint(fit_med)
# Call ci_i to find the bounds
out <- ci_i_one(1, npar = 5, which = "lbound",
            sem_out = fit_med, method = "wn",
            f_constr = fn_constr0)
out$bounds

}
\seealso{
\code{\link[=semlbci]{semlbci()}}, \code{\link[=ci_bound_wn_i]{ci_bound_wn_i()}}
}
