% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_free.R
\name{find_free}
\alias{find_free}
\title{Find the free parameters from an SEM output}
\usage{
find_free(sem_out)
}
\arguments{
\item{sem_out}{The SEM output. Currently supports \link[lavaan:lavaan-class]{lavaan::lavaan} outputs only.}
}
\value{
A boolean vector of the same length as the number of rows of the \code{lavaan}.
parameter tables. A position is \code{TRUE} if the corresponding parameter
is free.
}
\description{
Compute the standardized moderation effect given the \code{lm} output.
}
\details{
Currently supports \link[lavaan:lavaan-class]{lavaan::lavaan} outputs only.
}
\examples{
\dontrun{

data(cfa_two_factors)
library(lavaan)

mod <- 
"
f1 =~ x1 + x2 + a*x3
f2 =~ x4 + a*x5 + equal('f1=~x2')*x6
f1 ~~ 0*f2
asq := a^2
"

fit <- sem(mod, cfa_two_factors)
parameterTable(fit)
find_free(fit)

}
}
\keyword{internal}
