% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_variance_in_free.R
\name{find_variance_in_free}
\alias{find_variance_in_free}
\title{Find free variances in an SEM output}
\usage{
find_variance_in_free(sem_out)
}
\arguments{
\item{sem_out}{The SEM output. Currently \link[lavaan:lavaan-class]{lavaan::lavaan} outputs only.}
}
\value{
A boolean vector of the same length as the number of free parameters.
A position is \code{TRUE} if the corresponding free parameter
is a variance (op == "~~").
}
\description{
Find free variances in an SEM output
}
\details{
Currently supports \link[lavaan:lavaan-class]{lavaan::lavaan} outputs only.
}
\examples{
\dontrun{
cfa_two_factors

mod <-
"
f1 =~ x1 + x2 + x3
f2 =~ x4 + x5 + x6
"

fit <- lavaan::sem(mod, cfa_two_factors)
out <- find_variance_in_free(fit)
coef(fit)[out]
}
}
\keyword{internal}
