% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_bound_ur_i_helpers.R
\name{gen_sem_out_userp}
\alias{gen_sem_out_userp}
\title{Generate a Function to Fix a
User-Defined Parameter}
\usage{
gen_sem_out_userp(
  userp,
  sem_out,
  userp_name = "semlbciuserp1234",
  fix = TRUE,
  control_args = list(),
  iter.max = 10000,
  max_attempts = 5
)
}
\arguments{
\item{userp}{A function that receives
a \code{lavaan} object and returns a
scalar.}

\item{sem_out}{A \code{lavaan}-class object,
to which the function will be added
as a user-defined parameter.}

\item{userp_name}{The name of the
function \code{userp} to be used in the
\code{lavaan} model. It does not have to
be the name of the function in
\code{userp}. Should be changed only if it
conflicts with another object in the
parent environment, which should not
happen if the model is always fitted
in a clean R session.}

\item{fix}{If \code{TRUE}, the default,
the function generated is used to fix
the value of \code{userp} to a target
value using an equality constraint.
If \code{FALSE}, then the function simply
fits the model to the data.}

\item{control_args}{To be passed to
the argument of the same name in
\code{\link[lavaan:lavaan]{lavaan::lavaan()}}. Default is
\code{list()}. Can be used to set the
default values of this argument
in the generated function.}

\item{iter.max}{The maximum number of
iteration when the generated function
fit the model. Default is 10000.}

\item{max_attempts}{If the initial
fit with the equality constraint
fails, how many more attempts
will be made by the generated
function. Default is 5.}
}
\value{
If \code{fix} is \code{TRUE}, it returns a
function with these arguments:
\itemize{
\item \code{target}: The value to which the
user-defined parameter will be fixed
to.
\item \code{verbose}: If \code{TRUE}, additional
information will be printed when
fitting the model.
\item \code{control}: The values to be passed
as a list to the argument of the
same name in \code{\link[lavaan:lavaan]{lavaan::lavaan()}}.
\item \code{seed}: Numeric. If supplied, it
will be used in \code{\link[=set.seed]{set.seed()}} to
initialize the random number
generator. Necessary to reproduce
some results because random numbers
are used in some steps in \code{lavaan}.
If \code{NULL}, the default, \code{\link[=set.seed]{set.seed()}}
will not be called.
}

If \code{fix} is `FALSE, then it returns a
function with optional arguments that
will be ignored, Calling it will
simply fit the modified model to the
data. Useful for getting the value of
the user-defined parameter.
}
\description{
Add a function to a
\code{lavaan} object as a user-defined
parameter and return a function which
can fix this user-defined parameter
to a value.
}
\details{
This function is to be used internally
for generating a function for searching
a likelihood-based confidence bound.

It is exported because it needs to
be run in an external R process,
usually created by \code{callr} in other
internal functions.
}
\examples{

# TODO:
# - Add an example
data(simple_med)
dat <- simple_med


}
