% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_unique_name.R
\name{gen_unique_name}
\alias{gen_unique_name}
\title{Generate a unique name}
\usage{
gen_unique_name(source_names)
}
\arguments{
\item{source_names}{A vector of strings}
}
\value{
A string
}
\description{
Generate a unique name that does not duplicate with elements
in a vector of strings.
}
\details{
User supplies a vector of strings and then a name is generated that does
not duplicate with elements in this vector.

A helper function for generating names for new parameters.
}
\examples{

x <- c("A", "b", "xyz")

gen_unique_name(x)

}
\keyword{internal}
