% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_i_from_lor.R
\name{get_i_from_lor}
\alias{get_i_from_lor}
\title{Get the row number of a parameter from lhs, op, rhs, block, and group}
\usage{
get_i_from_lor(ptable, lor = NULL, lhs, op, rhs, block = NULL, group = NULL)
}
\arguments{
\item{ptable}{A lavaan parameter table (e.g., an object returned by
\code{\link[lavaan:lavParTable]{lavaan::parameterTable()}})}

\item{lor}{A one row data frame with these columns, lhs, op, rhs, block, and
group. lhs, op, and rhs are characters, and block and group are
numeric. block and group are optional. If they are omitted, an
error will be raised if there are more than one match.
If \code{NULL}, the data frame will be constructed from the following
arguments. Default is \code{NULL}.}

\item{lhs}{A string. The \code{lhs} in the parameter table.}

\item{op}{A string. The \code{op} in the parameter table.}

\item{rhs}{A string. The \code{rhs} in the parameter table.}

\item{block}{A number. The \code{block} in the parameter table. Optional}

\item{group}{A number. The \code{group} in the parameter table. Optional}
}
\value{
A vector of unique characters
}
\description{
Get the row number of a parameter in a parameter table
from lhs, op, rhs, block, and group
}
\details{
These columns will be extracted: \code{lhs}, \code{rhs}, \code{label}, and \code{plabel}.
}
\examples{

library(lavaan)

}
\keyword{internal}
