% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lhs_op_rhs.R
\name{get_lhs_op_rhs}
\alias{get_lhs_op_rhs}
\title{Get lhs, op, and rhs of a parameter from an SEM output}
\usage{
get_lhs_op_rhs(i, sem_out)
}
\arguments{
\item{i}{The position of the target parameters as in the parameter table of lavaan.}

\item{sem_out}{The SEM output. Currently \code{lavaan} output only.}
}
\value{
A one row data frame of lhs, op, and rhs of the parameter in the parameter table.
}
\description{
Get lhs, op, and rhs of a parameter from an SEM output
}
\details{
Currently supports \code{lavaan} output only.
}
\examples{
library(lavaan)
data(cfa_two_factors)
mod <- 
"
f1 =~ x1 + x2 + a*x3
f2 =~ x4 + a*x5 + equal('f1=~x2')*x6
f1 ~~ 0*f2
asq := a^2
"
fit <- sem(mod, cfa_two_factors)
}
