% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names_from_ptable.R
\name{get_names_from_ptable}
\alias{get_names_from_ptable}
\title{Get all the unique names from a parameter table}
\usage{
get_names_from_ptable(ptable)
}
\arguments{
\item{ptable}{A lavaan parameter table (e.g., an object returned by
\code{\link[lavaan:lavParTable]{lavaan::parameterTable()}})}
}
\value{
A vector of unique characters
}
\description{
Get all the unique names form a \link{lavaan} parameter table
}
\details{
These columns will be extracted: \code{lhs}, \code{rhs}, \code{label}, and \code{plabel}.
}
\examples{

library(lavaan)
mod <- 
"
y ~ x + m
z ~ a * y + m
"
ptable <- lavaanify(mod)
get_names_from_ptable(ptable)

}
\keyword{internal}
