% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_std_genfct.R
\name{get_std_genfct}
\alias{get_std_genfct}
\title{Generate a function to get the standardized estiamte of a parameter}
\usage{
get_std_genfct(fit, i)
}
\arguments{
\item{fit}{The source fit object.}

\item{i}{The position of the standardized parameter in the parameter table.}
}
\value{
A function to be used in a lavaan model.
}
\description{
Generate a function to get the standardized estiamte of a parameter
}
\examples{
# To Do

}
\keyword{internal}
