% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_std_i.R
\name{get_std_i}
\alias{get_std_i}
\title{Get standardized solution within lavaan}
\usage{
get_std_i(i, ...)
}
\arguments{
\item{i}{The positon of the parameter as in the parameter table.}

\item{...}{Optional arguments. Not used.}
}
\value{
The estimatate in the standardized solution.
}
\description{
Get standardized solution within lavaan
}
\examples{
# TODO

}
