% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cibound.R
\name{print.cibound}
\alias{print.cibound}
\title{Print diagnotic informantion of a confidence limit}
\usage{
\method{print}{cibound}(x, digits = 5, ...)
}
\arguments{
\item{x}{The output of ci_bound_xx_i function.}

\item{digits}{The number of digits after decimal point. To be passed to
\code{\link[=round]{round()}}. Default is 5.}

\item{...}{Other arguments. They will be ignored.}
}
\value{
Nothing
}
\description{
Print diagnotic informantion of a confidence limit
}
\details{
Print diagnotic informantion of a confidence limit
}
\examples{
# TODO

}
