% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reg_cor_near_up}
\alias{reg_cor_near_up}
\title{Generated from a regression model, with one correlation close to one.}
\format{
An object of class \code{data.frame} with 100 rows and 6 columns.
}
\usage{
reg_cor_near_up
}
\description{
Generated from a regression model, with one correlation close to one.
}
\keyword{datasets}
