% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_factor.R
\name{scaling_factor}
\alias{scaling_factor}
\title{Find the scaling factor used in satorra-2000 test}
\usage{
scaling_factor(sem_out, i, standardized = FALSE, pertubation_factor = 0.98)
}
\arguments{
\item{sem_out}{The source fit object.}

\item{i}{The position of the parameter as appeared in the
parameter table.}

\item{standardized}{If \code{TRUE}, the limit to be foound is for the standardized
solution. Default is \code{FALSE}.}

\item{pertubation_factor}{A factor to modify the original estimate. Default
is .98.}
}
\value{
The scaling factor
}
\description{
Find the scaling factor used in satorra-2000
test.
}
\examples{
# TODO

}
\keyword{internal}
