% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semlbci-package.R
\docType{package}
\name{semlbci-package}
\alias{semlbci-package}
\alias{_PACKAGE}
\title{semlbci: Likelihood Based Confidence Interval for SEM}
\description{
Find the likelihood based confidence intervals (LBCIs) for parameters in structural equation modeling, including standardized estimates such as correlatiosn and user-defined parameters such as indirect effect. Also support robust LBCIs.
}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

Other contributors:
\itemize{
  \item Ivan Jacob Agaloos Pesigan \email{i.j.a.pesigan@connect.um.edu.mo} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [contributor]
}

}
\keyword{internal}
