% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semlbci-package.R, R/semlbci.R
\docType{package}
\name{semlbci}
\alias{semlbci}
\title{semlbci: Likelihood Based Confidence Interval for SEM}
\usage{
semlbci(sem_out, ciperc = 0.95, ..., parallel = FALSE, ncpu = 2)
}
\arguments{
\item{sem_out}{The SEM output. Currently \code{lavaan} output only.}

\item{ciperc}{The proportion of coverage for the confidence interval. Default
is .95.}

\item{...}{Arguments to be passed to \code{ci_bound_i}.}

\item{parallel}{If \code{TRUE}, will use parallel.}

\item{ncpu}{The number of workers, if parallel is TRUE.}
}
\value{
The LBCIs for all free parameters
}
\description{
Find the LBCIs for all free parameters in an SEM output
}
\details{
Currently supports \code{lavaan} output only.
}
\examples{
library(lavaan)
data(cfa_two_factors)
mod <-
"
f1 =~ x1 + x2 + a*x3
f2 =~ x4 + a*x5 + equal('f1=~x2')*x6
f1 ~~ 0*f2
asq := a^2
"
fit <- sem(mod, cfa_two_factors)
}
