% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_constraint_nm.R
\name{set_constraint_nm}
\alias{set_constraint_nm}
\title{Set the constraint for finding the LBCI by the Neale-Miller-1997 approach}
\usage{
set_constraint_nm(
  i,
  sem_out,
  ciperc = 0.95,
  envir = NULL,
  get_fit_from_envir = FALSE
)
}
\arguments{
\item{i}{The position of the target parameters as in the parameter table of lavaan.}

\item{sem_out}{The SEM output. Currently \code{lavaan} output only.}

\item{ciperc}{The proportion of coverage for the confidence interval. Default
is .95.}

\item{envir}{The enviroment to stroe the state. Default is NULL}
}
\value{
A constraint function for nloptr.
}
\description{
Set the constraint for finding the LBCI by the Neale-Miller-1997 approach
}
\details{
Currently supports \code{lavaan} output only.
}
\examples{
library(lavaan)
data(cfa_two_factors)
mod <-
"
f1 =~ x1 + x2 + a*x3
f2 =~ x4 + a*x5 + equal('f1=~x2')*x6
f1 ~~ 0*f2
asq := a^2
"
fit <- sem(mod, cfa_two_factors)
#fn1 <- set_constraint(sem_out)

#fn1(coef(sem_out))
#fn1(runif(length(coef(sem_out))), .9, 1.1) * coef(sem_out))
#fn1(coef(sem_out) + 2)
#fn1(coef(sem_out) - .12)
}
