% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_start.R
\name{set_start}
\alias{set_start}
\title{Set the starting values for optimization}
\usage{
set_start(i = NULL, sem_out = NULL, which = NULL)
}
\arguments{
\item{i}{The position of the target parameters as in the parameter table of lavaan.}

\item{sem_out}{The SEM output. Currently \code{lavaan} output only.}

\item{which}{Whether the lower bound or the upper bound is to be found. Must be "lbound" or "ubound".}
}
\value{
A lavaan parameter table, with parameters estimated with target fixed to its
lower or upper Wald confidence limit.
}
\description{
Set the starting values for optimization
}
\details{
Currently supports \code{lavaan} output only.
}
\examples{
library(lavaan)
data(cfa_two_factors)
mod <- 
"
f1 =~ x1 + x2 + a*x3
f2 =~ x4 + a*x5 + equal('f1=~x2')*x6
f1 ~~ 0*f2
asq := a^2
"
fit <- sem(mod, cfa_two_factors)
}
