% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_start_wn.R
\name{set_start_wn}
\alias{set_start_wn}
\title{Set the starting values for optimization}
\usage{
set_start_wn(
  i = NULL,
  sem_out = NULL,
  which = NULL,
  standardized = FALSE,
  ciperc = 0.95
)
}
\arguments{
\item{i}{The position of the target parameters as in the parameter table of
lavaan.}

\item{sem_out}{The SEM output. Currently \code{lavaan} output only.}

\item{which}{Whether the lower bound or the upper bound is to be found.
Must be "lbound" or "ubound".}

\item{standardized}{If TRUE, the LBCI is for the standardized estimate.}

\item{ciperc}{The level of confidence for the confidence interval.}
}
\value{
A lavaan parameter table, with parameters estimated with target fixed to its
lower or upper Wald confidence limit.
}
\description{
Set the starting values for optimization
}
\details{
Currently supports \code{lavaan} output only.
}
\keyword{internal}
