% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_med}
\alias{simple_med}
\title{Dataset (Simple Mediation Model)}
\format{
A data frame with 200 rows and three variables:
\describe{
\item{x}{x, the independent variable}
\item{m}{m, the mediator}
\item{y}{y, the dependent variable}
}
}
\usage{
simple_med
}
\description{
Generated from a simple mediation model, n = 200
}
\details{
Supposed to be tested using this example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{library(lavaan)
mod <- "m ~ x
        y ~ m"
fit <- cfa(mod, simple_med)
summary(fit)
}\if{html}{\out{</div>}}
}
\examples{

print(head(simple_med), digits = 3)
nrow(simple_med)


}
\keyword{datasets}
