% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_constr.R
\name{test_constr}
\alias{test_constr}
\title{Test a limit using a constrained model}
\usage{
test_constr(
  fit,
  dat,
  ciperc = 0.95,
  parc,
  parc2 = "",
  modc0,
  ci_out,
  semfct,
  tol = 1e-04,
  update_args = list(),
  ...
)
}
\arguments{
\item{fit}{The source fit object.}

\item{dat}{The test data.}

\item{ciperc}{The level of confidence of the limit.}

\item{parc}{The character for the constraint(s).}

\item{parc2}{The character for the additional constraint(s).}

\item{modc0}{The base model to be constrained.}

\item{ci_out}{The output from the ci_bound_*_i function.}

\item{semfct}{The sem function to be used.}

\item{...}{Other arguments to be passed to the sem function}
}
\value{
The anova results
}
\description{
Test a limit using a constrained model
}
\examples{
# To

}
\keyword{internal}
