% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_model.R
\name{update_model}
\alias{update_model}
\title{Update a lavModel objec using user-supplied estimates}
\usage{
update_model(model, est)
}
\arguments{
\item{model}{A \code{lavModel} object to be updated.}

\item{est}{User supplied parameter estimates, as in the parameter table.}
}
\value{
An updated \code{lavModel} object.
}
\description{
Update a \code{lavaan} model using user-supplied estimates
}
\details{
Modified from the internal function \code{lav_model_set_parameters} in \code{lavaan}.

Obviously, supports \code{lavaan} output only.
}
\keyword{internal}
