% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfa_evar_near_zero}
\alias{cfa_evar_near_zero}
\title{Dataset: Two factors with one standardied error variance close to zero.}
\format{
A data frame with 100 rows and six variables:
\describe{
\item{x1}{x1}
\item{x2}{x2}
\item{x3}{x3}
\item{x4}{x4}
\item{x5}{x5}
\item{x6}{x6}
}
}
\usage{
cfa_evar_near_zero
}
\description{
Generated from a two-factor model, with one standardized error variacne
close to zero.
}
\examples{
\dontrun{
 # If fitted by the following model, the standardized
 # error variance of `x3` is close to zero.
 # Consequnently, the R-square of `x3` is close to one:

 library(lavaan)
 mod <- "f1 =~ x1 + x2 + x3
         f2 =~ x4 + x5 + x6"
 fit <- cfa(mod, cfa_evar_near_zero)
 summary(fit, standardized = TRUE, rsquare = TRUE)
}
}
\keyword{datasets}
