% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfa_two_factors}
\alias{cfa_two_factors}
\title{Dataset (CFA, Two Factors, Six Variables)}
\format{
A data frame with 500 rows and six variables, \code{x1} to \code{x6}.
}
\usage{
cfa_two_factors
}
\description{
Generated from a two-factor model with six variables, n = 500
}
\details{
This model is used for examples like this one:

\if{html}{\out{<div class="sourceCode">}}\preformatted{library(lavaan)
mod <- "f1 =~ x1 + x2 + x3
        f2 =~ x4 + x5 + x6"
fit <- cfa(mod, cfa_two_factors)
summary(fit)
}\if{html}{\out{</div>}}
}
\examples{

print(head(cfa_two_factors), digits = 3)
nrow(cfa_two_factors)

}
\keyword{datasets}
