% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfa_two_factors}
\alias{cfa_two_factors}
\title{Sample dataset for testing}
\format{
An object of class \code{data.frame} with 500 rows and 6 columns.
}
\usage{
cfa_two_factors
}
\description{
Generated from a two-factor model with six variables, n 500
}
\keyword{datasets}
