% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dependent.R
\name{find_dependent}
\alias{find_dependent}
\title{Find the free parameters on which a parameter depends on}
\usage{
find_dependent(i = NULL, sem_out = NULL, standardized = FALSE, signed = FALSE)
}
\arguments{
\item{i}{The position of the target parameter as in the parameter table of lavaan.}

\item{sem_out}{The SEM output. Currently \code{lavaan} output only.}

\item{standardized}{If TRUE, the LBCI is for the standardized estimate.}

\item{signed}{If TRUE, return a vector of 1 or -1 to indicate the direction of the dependence.
Default is \code{\link{FALSE}}.}
}
\value{
A numeric vector of the positions of the free parameters in the \code{lavaan} parameter table.
}
\description{
Find the free parameters on which a parameter depends on
}
\details{
Currently supports \code{lavaan} output only.
}
\examples{
library(lavaan)
data(cfa_two_factors)
mod <- 
"
f1 =~ x1 + x2 + a*x3
f2 =~ x4 + a*x5 + equal('f1=~x2')*x6
f1 ~~ 0*f2
asq := a^2
"
fit <- sem(mod, cfa_two_factors)
}
