% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_std.R
\name{get_std}
\alias{get_std}
\title{Get standardized estimate within lavaan}
\usage{
get_std(fit_str, i)
}
\arguments{
\item{fit_str}{A one-element character vector. The name of the fit object.}

\item{i}{The position of the parameter in the parameter table.}
}
\value{
The standardized estimate of the \emph{i} parameter.
}
\description{
This function is to be used to define user-defined
parameters in \code{lavaan}. Exported to be used by \code{\link[=scaling_factor]{scaling_factor()}}.
Not to be used by normal users.
}
\examples{
# TODO

}
