% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike_at.R
\name{plot.loglike_compare}
\alias{plot.loglike_compare}
\title{Plot the Output of 'loglike_compare()'}
\usage{
\method{plot}{loglike_compare}(x, y, type = "default", ...)
}
\arguments{
\item{x}{The output of \code{\link[=loglike_compare]{loglike_compare()}}.}

\item{y}{Not used.}

\item{type}{Character. If \code{"ggplot2"}, will use \link{ggplot2} to plot the graph.
Default is \code{"default"} and R base graphics is used.}

\item{...}{Optional arguments. Ignored.}
}
\value{
...
}
\description{
...
}
\details{
...
}
\examples{

library(lavaan)
data(simple_med)
mod <-
"
m ~ a*x
y ~ b*m
ab:= a*b
asq:= a^2
"
fit_med <- sem(mod, simple_med, fixed.x = FALSE)
p_table <- parameterTable(fit_med)

pars <- c("m ~ x",
          "y ~ m",
          "asq := 1",
          "ab  := 2",
          "not in table")
out <- syntax_to_i(pars, fit_med)
out
p_table[out, ]
}
