% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike_at.R
\name{plot.loglike_compare}
\alias{plot.loglike_compare}
\title{Plot the Output of 'loglike_compare()'}
\usage{
\method{plot}{loglike_compare}(x, y, type = "default", add_pvalues = FALSE, ...)
}
\arguments{
\item{x}{The output of \code{\link[=loglike_compare]{loglike_compare()}}.}

\item{y}{Not used.}

\item{type}{Character. If \code{"ggplot2"}, will use \link{ggplot2} to plot
the graph. Default is \code{"default"} and R base graphics
is used.}

\item{add_pvalues}{If \code{TRUE}, likelihood ratio test \emph{p}-values will be
plot for the confidence limits.}

\item{...}{Optional arguments. Ignored.}
}
\value{
Nothing if \code{type = "default"}, the generated \link{ggplot2}
graph if \code{type = "ggplot2"}.
}
\description{
Visualize the likelihood of a model with a parameter
fixed to values in a range.
}
\details{
Given the output of \code{\link[=loglike_compare]{loglike_compare()}}, it plots the
log-likelihood based on quadratic approximation and that
based on the "true" log-likelihood. The log-likelihood is
scaled to have a maximum of zero as suggested by Pawitan
(2013).
}
\examples{

library(lavaan)
data(simple_med)
mod <-
"
m ~ a*x
y ~ b*m
ab:= a*b
asq:= a^2
"
fit_med <- sem(mod, simple_med, fixed.x = FALSE)
p_table <- parameterTable(fit_med)

pars <- c("m ~ x",
          "y ~ m",
          "asq := 1",
          "ab  := 2",
          "not in table")
out <- syntax_to_i(pars, fit_med)
out
p_table[out, ]
}
\references{
Pawitan, Y. (2013). \emph{In all likelihood: Statistical
modelling and inference using likelihood}. Oxford University Press.
}
