% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_semlbci.R
\name{print.semlbci}
\alias{print.semlbci}
\title{Print the results of semlbci}
\usage{
\method{print}{semlbci}(
  x,
  digits = 3,
  annotation = TRUE,
  time = FALSE,
  verbose_if_needed = TRUE,
  drop_no_lbci = TRUE,
  ...
)
}
\arguments{
\item{x}{The output of \code{\link[=semlbci]{semlbci()}}.}

\item{digits}{The number of digits after decimal point. To be passed to
\code{\link[=round]{round()}}. Default is 3.}

\item{annotation}{If \code{TRUE}, print table notes. Default is \code{TRUE}.}

\item{time}{If \code{TRUE}, print the time spent on each limit. Default is
\code{FALSE}.}

\item{verbose_if_needed}{If \code{TRUE}, additional diagnostic information will
be printed if necessary. If \code{FALSE}, additional
diagnostic information will always be printed.
Default is \code{TRUE}.}

\item{drop_no_lbci}{If \code{TRUE}, parameters not included in the search will
be removed. Default is \code{TRUE}.}

\item{...}{Other arguments. They will be ignored.}
}
\value{
Nothing
}
\description{
Print the results of semlbci
}
\details{
Print the results of semlbci
}
\examples{
# TODO

}
