% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reg_cor_near_one}
\alias{reg_cor_near_one}
\title{Dataset (Six Variables, One Correlation Close to One)}
\format{
A data frame with 100 rows and six variables:
\describe{
\item{x1}{x1}
\item{x2}{x2}
\item{x3}{x3}
\item{x4}{x4, with correlation with x5 nearly equal to 1}
\item{x5}{x5, with correlation with x4 nearly equal to 1}
\item{y}{y, the dependent variable}
}
}
\usage{
reg_cor_near_one
}
\description{
Generated from a regression model six variables,
x4~~x5 correlation close to one.
}
\details{
Supposed to be tested using this example:\preformatted{out <- lm(y ~ x1 + x2 + x3 + x4 + x5, reg_cor_near_one)
summary(out)
cor(reg_cor_near_one[, c("x4", "x5")])
}
}
\keyword{datasets}
