% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semlbci-package.R
\docType{package}
\name{semlbci-package}
\alias{semlbci-package}
\alias{_PACKAGE}
\title{semlbci: Likelihood-Based Confidence Interval in Structural Equation Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Forms likelihood-based confidence intervals (LBCIs) for parameters in structural equation modeling, introduced in Cheung and Pesigan (2023) \doi{10.1080/10705511.2023.2183860}. Currently implements the algorithm illustrated by Pek and Wu (2018) \doi{10.1037/met0000163}, and supports the robust LBCI proposed by Falk (2018) \doi{10.1080/10705511.2017.1367254}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sfcheung.github.io/semlbci/}
  \item Report bugs at \url{https://github.com/sfcheung/semlbci/issues}
}

}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

Other contributors:
\itemize{
  \item Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [contributor]
}

}
\keyword{internal}
