% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semlbci-package.R
\docType{package}
\name{semlbci-package}
\alias{semlbci-package}
\alias{_PACKAGE}
\title{semlbci: Likelihood Based Confidence Interval for SEM}
\description{
Functions for forming the likelihood-based confidence intervals (LBCIs) for parameters in structural equation modeling. Currently implements the algorithm illustrated by Pek and Wu (2018) \doi{10.1037/met0000163}, and supports the robust LBCI proposed by Falk (2018) \doi{10.1080/10705511.2017.1367254}.
}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

Other contributors:
\itemize{
  \item Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [contributor]
}

}
\keyword{internal}
