% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semlbci-package.R, R/semlbci.R
\docType{package}
\name{semlbci}
\alias{semlbci}
\title{semlbci: Likelihood Based Confidence Interval for SEM}
\usage{
semlbci(
  sem_out,
  pars = NULL,
  ciperc = 0.95,
  standardized = FALSE,
  method = "wn",
  ...,
  parallel = FALSE,
  ncpu = 2
)
}
\arguments{
\item{sem_out}{The SEM output. Currently \code{lavaan} output only.}

\item{pars}{Positions of the parameters for which the LBCI to be found.
Use the position as appeared on the parameter tables of the fit object.
If NULL, the default, then LBCIs will be found for all free parameters.
Can also be a vector of strings to indicate the paramters on the
parameter table.}

\item{ciperc}{The proportion of coverage for the confidence interval. Default
is .95.}

\item{standardized}{If TRUE, the LBCI is for the standardized estimate.}

\item{method}{The approach to be used. Can be "wn" (Wu-Neale-2012) or "nm"
(Neale-Miller-1997). Default is "wn".}

\item{...}{Arguments to be passed to \code{ci_bound_i}.}

\item{parallel}{If \code{TRUE}, will use parallel. Currently disabled.
Need to find out how to make \code{lavaan::udpate} works
currently in the workers.}

\item{ncpu}{The number of workers, if parallel is TRUE.}
}
\value{
The LBCIs for all free parameters
}
\description{
Find the LBCIs for all free parameters in an SEM output
}
\details{
Currently supports \code{lavaan} output only.
}
\examples{
library(lavaan)
data(cfa_two_factors)
mod <-
"
f1 =~ x1 + x2 + a*x3
f2 =~ x4 + a*x5 + equal('f1=~x2')*x6
f1 ~~ 0*f2
asq := a^2
"
fit <- sem(mod, cfa_two_factors)
}
