% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semlbci.R
\name{semlbci}
\alias{semlbci}
\title{Find the LBCIs for selected free parameters in an SEM output}
\usage{
semlbci(
  sem_out,
  pars = NULL,
  ciperc = 0.95,
  standardized = FALSE,
  method = "nm",
  ...,
  parallel = FALSE,
  ncpu = 2
)
}
\arguments{
\item{sem_out}{The SEM output. Currently supports a \link[lavaan:lavaan-class]{lavaan::lavaan}
output only.}

\item{pars}{The positions of the parameters for which the LBCI to be found.
Use the position as appeared on the parameter tables of the fit object.
If NULL, the default, then LBCIs will be found for all free parameters.
Can also be a vector of strings to indicate the paramters on the
parameter table.}

\item{ciperc}{The proportion of coverage for the confidence interval. Default
is .95.}

\item{standardized}{If TRUE, the LBCI is for the standardized estimate.}

\item{method}{The approach to be used. Can be "wn" (Wu-Neale-2012) or "nm"
(Neale-Miller-1997). Default is "nm".}

\item{...}{Arguments to be passed to \code{ci_bound_i}.}

\item{parallel}{If \code{TRUE}, will use parallel. Currently disabled.
Need to find out how to make \code{lavaan::udpate} works
currently in the workers.}

\item{ncpu}{The number of workers, if parallel is TRUE.}
}
\value{
A modified \code{lavaan} parameter table with the LBCIs for selected
free parameters
}
\description{
Find the likelihood-based confidence intervals (LBCIs) for
selected free parameters in an SEM output
}
\details{
This function finds the positions of the selected parameters in the parameter
table and then call \code{\link[=ci_i]{ci_i()}} once for each of them. Please see \code{\link[=ci_i]{ci_i()}} and
it child functions (currently \code{\link[=ci_bound_nm_i]{ci_bound_nm_i()}} and \code{\link[=ci_bound_i]{ci_bound_i()}}) for the
technical details.

The SEM output will first checked by \code{\link[=check_sme_out]{check_sme_out()}} to see whether the model
and the estimation method are supported.

Currently supports \link[lavaan:lavaan-class]{lavaan::lavaan} outputs only.
}
\examples{
library(lavaan)
data(cfa_two_factors)
mod <-
"
f1 =~ x1 + x2 + a*x3
f2 =~ x4 + a*x5 + equal('f1=~x2')*x6
f1 ~~ 0*f2
asq := a^2
"
fit <- sem(mod, cfa_two_factors)
}
\references{
Neale, M. C., & Miller, M. B. (1997). The use of likelihood-based confidence
intervals in genetic models. \emph{Behavior Genetics, 27}(2), 113–120.
\url{https://doi.org/10.1023/A:1025681223921}

Pritikin, J. N., Rappaport, L. M., & Neale, M. C. (2017). Likelihood-based
confidence intervals for a parameter with an upper or lower bound.
\emph{Structural Equation Modeling: A Multidisciplinary Journal, 24}(3), 395–401.
\url{https://doi.org/10.1080/10705511.2016.1275969}
}
