% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semlbci.R
\name{semlbci}
\alias{semlbci}
\title{Find the LBCIs for selected free parameters in an SEM output}
\usage{
semlbci(
  sem_out,
  pars = NULL,
  ciperc = 0.95,
  standardized = FALSE,
  method = "wn",
  ...,
  parallel = FALSE,
  ncpu = 2
)
}
\arguments{
\item{sem_out}{The SEM output. Currently supports \link[lavaan:lavaan-class]{lavaan::lavaan}
outputs only.}

\item{pars}{The positions of the parameters for which the LBCI to be found.
Use the position as appeared on the parameter tables of the
\code{sem_out}. If NULL, the default, then LBCIs for all free
parameters will be searched.
Can also be a vector of strings to indicate the paramters
on the parameter table. The parameters should be specified in
\code{\link[lavaan:lavaan]{lavaan::lavaan()}} syntax. The vector of strings will be
converted by \code{\link[=syntax_to_i]{syntax_to_i()}} to parameter positions.}

\item{ciperc}{The proportion of coverage for the confidence interval. Default
is .95, requesting a 95 percent confidence interval.}

\item{standardized}{If \code{TRUE}, the LBCI is for the standardized estimates.}

\item{method}{The approach to be used to search for the confidence limits.
Currently only "wn" (Wu-Neale-2012) is supported.}

\item{...}{Arguments to be passed to \code{\link[=ci_bound_wn_i]{ci_bound_wn_i()}}.}

\item{parallel}{If \code{TRUE}, will use \link{parallel}. Currently disabled and
so this argument will be ignored.}

\item{ncpu}{The number of workers, if \code{parallel} is \code{TRUE}. Currently
ignored.}
}
\value{
A parameter table generated by \code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}},
with the LBCIs for selected parameters added. Diagnostic information,
if requested, will be included in the attributes
}
\description{
Find the likelihood-based confidence intervals (LBCIs) for
selected free parameters in an SEM output
}
\details{
This function finds the positions of the selected parameters in the parameter
table and then call \code{\link[=ci_i]{ci_i()}} once for each of them. Please see \code{\link[=ci_i]{ci_i()}} and
the functions it call to find a confidence limit (currently
\code{\link[=ci_bound_wn_i]{ci_bound_wn_i()}}) for the technical details.

The SEM output will first checked by \code{\link[=check_sem_out]{check_sem_out()}} to see whether the
model and the estimation method are supported.

Currently supports \link[lavaan:lavaan-class]{lavaan::lavaan} outputs only.
}
\examples{
\dontrun{
#library(lavaan)
#mod <-
#"
#m ~ a*x
#y ~ b*m
#ab := a * b
#"
#fit_med <- sem(mod, simple_med, fixed.x = FALSE)
#p_table <- parameterTable(fit_med)
#p_table
#lbci_med <- semlbci(fit_med,
#                    pars = c("m ~ x",
#                             "y ~ m",
#                             "ab :="),
#                    method = "wn")
#lbci_med
}
}
\references{
Neale, M. C., & Miller, M. B. (1997). The use of likelihood-based confidence
intervals in genetic models. \emph{Behavior Genetics, 27}(2), 113–120.
\url{https://doi.org/10.1023/A:1025681223921}

Pritikin, J. N., Rappaport, L. M., & Neale, M. C. (2017). Likelihood-based
confidence intervals for a parameter with an upper or lower bound.
\emph{Structural Equation Modeling: A Multidisciplinary Journal, 24}(3), 395–401.
\url{https://doi.org/10.1080/10705511.2016.1275969}
}
