% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semlbci.R
\name{semlbci}
\alias{semlbci}
\title{Likelihood-Based Confidence Interval}
\usage{
semlbci(
  sem_out,
  pars = NULL,
  include_user_pars = TRUE,
  include_variances = FALSE,
  ciperc = 0.95,
  standardized = FALSE,
  method = "wn",
  robust = "none",
  try_k_more_times = 2,
  ...,
  parallel = FALSE,
  ncpus = 2,
  use_pbapply = TRUE
)
}
\arguments{
\item{sem_out}{The SEM output. Currently supports
\link[lavaan:lavaan-class]{lavaan::lavaan} outputs only.}

\item{pars}{The positions of the parameters for which the LBCIs are
to be searched. Use the position as appeared on the parameter
tables of the \code{sem_out}. If \code{NULL}, the default, then LBCIs for
all free parameters will be searched. Can also be a vector of
strings to indicate the parameters on the parameter table. The
parameters should be specified in \code{\link[lavaan:lavaan]{lavaan::lavaan()}} syntax. The
vector of strings will be converted by \code{\link[=syntax_to_i]{syntax_to_i()}} to
parameter positions.}

\item{include_user_pars}{Logical. Whether all user-defined parameters
are automatically included when \code{pars} is not set. Default is \code{TRUE}.
If \code{pars} is explicitly set, this argument will be ignored.}

\item{ciperc}{The proportion of coverage for the confidence
interval. Default is .95, requesting a 95 percent confidence
interval.}

\item{standardized}{If \code{TRUE}, the LBCI is for the standardized estimates.}

\item{method}{The approach to be used to search for the confidence
limits. Currently only \code{"wn"} (Wu-Neale-2012), the default, is
supported.}

\item{robust}{Whether the LBCI based on robust likelihood ratio
test is to be found. Only \code{"satorra.2000"} in \code{\link[lavaan:lavTestLRT]{lavaan::lavTestLRT()}}
is supported for now, implemented by the method proposed by Falk
(2018). If \code{"none"}, the default, then likelihood ratio test based
on maximum likelihood estimation will be used.}

\item{try_k_more_times}{How many more times to try if failed.
Default is 2.}

\item{...}{Arguments to be passed to \code{\link[=ci_bound_wn_i]{ci_bound_wn_i()}}.}

\item{parallel}{If \code{TRUE}, will use \code{parallel} to parallelize the search.}

\item{ncpus}{The number of workers, if \code{parallel} is \code{TRUE}. Default is 2.}

\item{use_pbapply}{If \code{TRUE}, \code{parallel} is \code{TRUE}, and \code{pbapply}
is installed, \code{\link[pbapply:pbapply]{pbapply::pbapply()}} will be used to display a
progress bar when finding the intervals. Default is \code{TRUE}.
Ignored if \code{parallel} is \code{FALSE}.}

\item{remove_variances}{Logical. Whether variances and error variances
will be removed. Default is \code{TRUE}, removing all variances and error
variances even if specified in \code{pars}.}
}
\value{
A parameter table generated by
\code{\link[lavaan:parameterEstimates]{lavaan::parameterEstimates()}}, with the LBCIs for selected
parameters added. Diagnostic information, if requested, will be
included in the attributes
}
\description{
Finds the likelihood-based confidence intervals (LBCIs) for
selected free parameters in an SEM output.
}
\details{
\code{\link[=semlbci]{semlbci()}} finds the positions of the selected parameters
in the parameter table and then calls \code{\link[=ci_i_one]{ci_i_one()}} once for each of
them. Please see \code{\link[=ci_i_one]{ci_i_one()}} and the functions it calls to find a
confidence limit (currently \code{\link[=ci_bound_wn_i]{ci_bound_wn_i()}}) for the technical
details.

The SEM output will first be checked by \code{\link[=check_sem_out]{check_sem_out()}} to see
whether the model and the estimation method are supported.

Currently supports \link[lavaan:lavaan-class]{lavaan::lavaan} outputs only.
}
\examples{

library(lavaan)
mod <-
"
m ~ a*x
y ~ b*m
ab := a * b
"
fit_med <- sem(mod, simple_med, fixed.x = FALSE)
p_table <- parameterTable(fit_med)
p_table
lbci_med <- semlbci(fit_med,
                    pars = c("m ~ x",
                             "y ~ m",
                             "ab :="))
lbci_med

}
\references{
Falk, C. F. (2018). Are robust standard errors the best approach
for interval estimation with nonnormal data in structural equation
modeling? \emph{Structural Equation Modeling: A Multidisciplinary
Journal, 25}(2), 244-266.
\doi{10.1080/10705511.2017.1367254}

Neale, M. C., & Miller, M. B. (1997). The use of likelihood-based confidence
intervals in genetic models. \emph{Behavior Genetics, 27}(2), 113-120.
\doi{10.1023/A:1025681223921}

Pritikin, J. N., Rappaport, L. M., & Neale, M. C. (2017). Likelihood-based
confidence intervals for a parameter with an upper or lower bound.
\emph{Structural Equation Modeling: A Multidisciplinary Journal, 24}(3), 395-401.
\doi{10.1080/10705511.2016.1275969}
}
