% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_constraint.R
\name{set_constraint}
\alias{set_constraint}
\title{Set the constraint for finding the LBCI}
\usage{
set_constraint(sem_out, ciperc = 0.95)
}
\arguments{
\item{sem_out}{The SEM output. Currently \code{lavaan} output only.}

\item{ciperc}{The proportion of coverage for the confidence interval. Default
is .95.}
}
\value{
A constraint function for nloptr.
}
\description{
Set the constraint for finding the LBCI
}
\details{
Currently supports \code{lavaan} output only.
}
\examples{
library(lavaan)
data(cfa_two_factors)
mod <-
"
f1 =~ x1 + x2 + a*x3
f2 =~ x4 + a*x5 + equal('f1=~x2')*x6
f1 ~~ 0*f2
asq := a^2
"
fit <- sem(mod, cfa_two_factors)
#fn1 <- set_constraint(sem_out)

#fn1(coef(sem_out))
#fn1(runif(length(coef(sem_out))), .9, 1.1) * coef(sem_out))
#fn1(coef(sem_out) + 2)
#fn1(coef(sem_out) - .12)
}
