% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_constraint.R
\name{set_constraint}
\alias{set_constraint}
\title{Constraint for Finding the LBCI by the Wu-Neale-2012 Approach}
\usage{
set_constraint(sem_out, ciperc = 0.95)
}
\arguments{
\item{sem_out}{The SEM output. Currently supports
\link[lavaan:lavaan-class]{lavaan::lavaan} outputs only.}

\item{ciperc}{The probability of coverage for the confidence
interval. Default is .95.}
}
\value{
A constraint function for \link{nloptr}.
}
\description{
Sets the constraint for finding the likelihood-based
confidence interval by the Wu-Neale-2012 approach.
}
\details{
The Wu-Neale-2012 approach uses a simple objective function that is
optimized with a constraint. \link{set_constraint} generates the
constraint function used by \code{\link[=ci_bound_wn_i]{ci_bound_wn_i()}}.

Currently supports \link[lavaan:lavaan-class]{lavaan::lavaan} outputs only.

This function is not to be used by normal users.
}
\examples{

library(lavaan)
data(simple_med)
dat <- simple_med
mod <-
"
m ~ x
y ~ m
"
fit_med <- sem(mod, simple_med, fixed.x = FALSE)

fn_constr0 <- set_constraint(fit_med)
out <- fn_constr0(coef(fit_med), sem_out = fit_med)
out
lavTech(fit_med, "optim")$fx

}
