% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_med}
\alias{simple_med}
\title{Generated from a simple mediation model, n 500}
\format{
An object of class \code{data.frame} with 500 rows and 3 columns.
}
\usage{
simple_med
}
\description{
Generated from a simple mediation model, n 500
}
\keyword{datasets}
