% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_med}
\alias{simple_med}
\title{Dataset (Simple Mediation Model)}
\format{
A data frame with 500 rows and three variables:
\describe{
\item{x}{x, the independent variable}
\item{m}{m, the mediator}
\item{y}{y, the dependent variable}
}
}
\usage{
simple_med
}
\description{
Generated from a simple mediation model, n = 500
}
\details{
Supposed to be tested using this example:\preformatted{library(lavaan)
mod <- "m ~ x
        y ~ m"
fit <- cfa(mod, simple_med)
summary(fit)
}
}
\keyword{datasets}
