% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax_to_i.R
\name{syntax_to_i}
\alias{syntax_to_i}
\title{Convert lavaan syntax to positions in the fit object parameter table}
\usage{
syntax_to_i(syntax, sem_out)
}
\arguments{
\item{syntax}{A vector of parameters, defined as in lavaan.}

\item{sem_out}{The SEM output. Currently \code{lavaan} output only.}
}
\value{
A vector of positions in the parameter table.
}
\description{
Convert lavaan syntax to positions in the fit object parameter table
}
\details{
Currently supports \code{lavaan} output only.
}
\examples{
library(lavaan)
data(cfa_two_factors)
mod <- 
"
f1 =~ x1 + x2 + a*x3
f2 =~ x4 + a*x5 + equal('f1=~x2')*x6
f1 ~~ 0*f2
asq := a^2
"
fit <- sem(mod, cfa_two_factors)
}
