% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_model.R
\name{update_model}
\alias{update_model}
\title{Update a lavModel objec using user-supplied estimates}
\usage{
update_model(model, est)
}
\arguments{
\item{model}{A lavModel object to be updated.}

\item{est}{User supplied parameter estimates, as in the parameter table.}
}
\value{
An updated lavModel object.
}
\description{

}
\details{
Modified from lav_model_set_parameters in lavaan.

Obviously, supports \code{lavaan} output only.
}
\examples{
library(lavaan)
data(cfa_two_factors)
mod <- 
"
f1 =~ x1 + x2 + a*x3
f2 =~ x4 + a*x5 + equal('f1=~x2')*x6
f1 ~~ 0*f2
asq := a^2
"
fit <- sem(mod, cfa_two_factors)
}
